CMakeLists.txt: remove the distribution check
Summary:
The CMakeLists.txt file was checking the distribution on the system it
was running on. It was bailing out if it wasn't Fedora, Debian or
CentOS.

The DISTRIB variable set by this check is then ignored by the rest of
the file. Let's get rid of this check so that LAO can be build on more
systems.

Test Plan: CI

Reviewers: jvillette, bddinechin

Maniphest Tasks: T17869

Differential Revision: https://phab.kalray.eu/D17241

(cherry picked from commit 2f42ce7fc04d2ceac9c9828e042bbc0fcb5627fe)